+++ chrome/browser/ui/app_list/search/util/mrfu_cache.cc
@@ -9,6 +9,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "base/files/file_path.h"
 #include "base/logging.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/time/time.h"
 
 namespace app_list {
@@ -207,7 +208,7 @@ void MrfuCache::Decay(Score* score) {
 }
 
 void MrfuCache::MaybeCleanup() {
-  if (proto_->items_size() < 2u * max_items_)
+  if (base::checked_cast<size_t>(proto_->items_size()) < 2u * max_items_)
     return;
 
   // Ensure all scores are up to date, and then keep all those over the
